Original alignment approach GUI required
An original approach for analyzing the alignment of fibronectin fibers using the OrientationJ plugin through the graphical interface of ImageJ, compatible with both Windows and macOS.
To run the script, Fiji and the OrientationJ plugin must be manually installed in the same folder as the script.
The program operates within a Python environment that can be installed using the environment_uma_original.yaml file.
It supports input folders containing .tiff, .nd2, .oif, and .oib files.